Branch3
In this branch, I decided to remove all plots/figures except the plot containing the distribution of posteriors and a graph that plotted the chosen model runs on top of the data. I did this because it would take a very long time and heat up my computer significantly when I would run the 50K models if I had all the previous plots included. I also moved some constants that were inside the for loop (recreating vectors and data frames that really only needed to be made once )and moved them outside, which speed up the computation.